Changes Updated

+++ day-6/src/main/java/com/restapiday6/day6/social/controller/UserController.java
@@ -2,7 +2,7 @@ package com.restapiday6.day6.social.controller;
 
 import com.restapiday6.day6.social.model.Users;
 import com.restapiday6.day6.social.services.UsersServices;
-import com.restapiday6.day6.social.exception.userNotFoundException;
+import com.restapiday6.day6.social.exception.UserNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
 @RestController
-public class UserResource {
+public class UserController {
     @Autowired
     public UsersServices service;
 
@@ -24,22 +24,21 @@ public class UserResource {
     public Users findOne(@PathVariable int id){
        Users user=service.findById(id);
        if(user== null){
-           throw  new userNotFoundException("User not found with Id:"+id);
+           throw  new UserNotFoundException("User not found with Id:"+id);
        }
        return user;
     }
-
     @DeleteMapping("/users/{id}")
     public void deleteOne(@PathVariable int id){
-        Users user=service.deleteById(id);
+        Users users= service.findById(id);
+        if(users== null){
+            throw  new UserNotFoundException("User not found with Id:"+id);
+        }
+        service.deleteById(id);
     }
     @PostMapping("/users")
     public ResponseEntity<Users> createUse(@Validated @RequestBody Users user){
         Users saveduser= service.saveUser(user);
-//        URI location= ServletUriComponentsBuilder.fromCurrentRequest()
-//                .path("/{id}")
-//                .buildAndExpand(saveduser.getId())
-//                .toUri();
         return ResponseEntity.created(null).body(saveduser);
     }
